perf: 不在获取数据时做 schema 转换
